Use new `@deprecated` decorator
The `@deprecated` decorator has been proposed in PEP 702 and is part of Python 3.13.
The huge benefit of the decorator is that it moves deprecations into the type system, allowing type checkers to catch calls to these methods. Runtime `DeprecationWarnings` are raised automatically and only once on first use.
As usual, the `@deprecated` decorator is backported to older versions of Python in `typing_extensions`. We can use the native implementation `warnings.deprecated` once Python 3.13+ is the minimum version.