Use `Self` return type where applicable
The `Self` return type has been introduced in Python 3.11.
It allows to express that the return type is of the current enclosed class while correctly considering subclasses.

As usual, the `Self` type is backported to older versions of Python in `typing-extensions`. We can use the native implementation `typing.Self` once Python 3.11+ is the minimum version.